+++ src/Stack/Options/GhciParser.hs
@@ -8,6 +8,7 @@ import           Options.Applicative.Builder.Extra
 import           Paths_stack                       as Meta
 import           Stack.Config                      (packagesParser)
 import           Stack.Ghci                        (GhciOpts (..))
+import           Stack.Options.BuildParser         (flagsParser)
 
 -- | Parser for GHCI options
 ghciOptsParser :: Parser GhciOpts
@@ -23,6 +24,12 @@ ghciOptsParser = GhciOpts
              <*> fmap concat (many (argsOption (long "ghci-options" <>
                                     metavar "OPTION" <>
                                     help "Additional options passed to GHCi")))
+             <*> many
+                     (textOption
+                          (long "ghc-options" <>
+                           metavar "OPTION" <>
+                           help "Additional options passed to both GHC and GHCi"))
+             <*> flagsParser
              <*> optional
                      (strOption (long "with-ghc" <>
                                  metavar "GHC" <>